# swagger.yaml
openapi: 3.0.3
info:
  title: Transaction API
  description: |
    Personal finance transaction API built with Actix-Web (Rust) and SQLite.
    Manages sources (wallets/accounts), earnings, spendings, and debts.
    App settings live in login-api as of 2026-07-27.
    
    All user-specific endpoints are scoped under `/api/user` and require a bearer JWT.
    The validated token's `sub` claim is used as the `created_by` user identifier.
  version: 2.0.0
  contact:
    name: Transaction API
servers:
  - url: http://127.0.0.1:3000
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /api/user/spendings/{spending_id}/details:
    get:
      tags: [Spendings]
      summary: Get the line items of a single spending
      operationId: getSpendingDetailsById
      parameters:
        - name: spending_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Spending details retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseSpendingDetails"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/user/spending-details/{spending_detail_id}/checked:
    put:
      tags: [Spendings]
      summary: Tick or untick one line item
      description: >
        The spending header is immutable once saved, so this is the only field
        of a stored breakdown a client can change. Used by the checklist on the
        Flutter transaction detail screen.
      operationId: putSpendingDetailChecked
      parameters:
        - name: spending_detail_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [checked]
              properties:
                checked:
                  type: boolean
                  example: false
      responses:
        "200":
          description: Line item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseResponse"
        "400":
          description: Malformed spending detail id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "404":
          description: No line item with that id belongs to the caller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Spending Categories ───────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ResponseError"
  schemas:
...
    BaseResponse:
      type: object
      properties:
        status:
          type: string
          example: "Success"
        code:
          type: integer
          example: 200
        message:
          type: string
        description:
          type: string
        success:
          type: boolean
        data:
          nullable: true
...
    ResponseError:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            status:
              example: "Error"
            success:
              example: false
            data:
              nullable: true
              example: null

    # ── App Settings schemas moved to login-api/swagger.yaml ────────────────

    # ── Sources ─────────────────────────────────────────────────────────────
...
    SpendingDetailInput:
      type: object
      required: [item_name]
      properties:
        item_name:
          type: string
          example: "Indomie Goreng"
        quantity:
          type: number
          format: double
          default: 1
          example: 2
        unit_price:
          type: number
          format: double
          example: 3500
        amount:
          type: number
          format: double
          description: Line total. Falls back to `quantity * unit_price` when 0.
          example: 7000
        note:
          type: string
          example: ""
        checked:
          type: boolean
          default: true
          description: >
            Whether the item is ticked on the client's checklist. Unticked items
            stay in the breakdown but are left out of the derived total.
    SpendingDetail:
      allOf:
        - $ref: "#/components/schemas/SpendingDetailInput"
        - type: object
          properties:
            spending_detail_id:
              type: string
              format: uuid
            spending_id:
              type: string
              format: uuid
            is_checked:
              type: boolean
              example: true
            created_date:
              type: string
              format: date-time
            created_by:
              type: string
            is_active:
              type: integer
              example: 1
...
    ResponseSpendingDetails:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/SpendingDetail"
...